// src/apis/models/Volo_Abp_OpenIddict_Scopes_Dtos_CreateScopeInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Volo_Abp_OpenIddict_Scopes_Dtos_CreateScopeInput = {
    readonly extraProperties?: Record<string, any> | null;
    name: string;
    displayName?: string | null;
    description?: string | null;
    resources?: Array<string> | null;
};